@@ -5,19 +5,21 @@
   "type": "module",
   "scripts": {
     "dev": "vite",
-    "build": "vue-tsc && vite build",
+    "build": "node build.js",
     "preview": "vite preview",
     "test:e2e": "cypress run",
-    "test:unit": "vitest",
-    "lint": "eslint ."
+    "e2e:open": "cypress open",
+    "lint": "eslint .",
+    "format": "prettier --write ."
   },
   "dependencies": {
     "@capacitor/android": "6.1.2",
     "@capacitor/app": "6.0.1",
-    "@capacitor/core": "6.1.2",
+    "@capacitor/core": "^6.1.2",
     "@capacitor/haptics": "6.0.1",
     "@capacitor/ios": "6.1.2",
     "@capacitor/keyboard": "6.0.2",
+    "@capacitor/preferences": "^6.0.2",
     "@capacitor/status-bar": "6.0.1",
     "@ionic/vue": "^8.0.0",
     "@ionic/vue-router": "^8.0.0",
@@ -34,21 +36,26 @@
     "vuex": "^4.0.2"
   },
   "devDependencies": {
-    "@capacitor/cli": "6.1.2",
+    "@capacitor/cli": "^6.1.2",
+    "@testing-library/jest-dom": "^6.6.2",
+    "@testing-library/vue": "^6.4.2",
+    "@typescript-eslint/eslint-plugin": "^8.12.2",
+    "@typescript-eslint/parser": "^8.12.2",
     "@vitejs/plugin-legacy": "^5.0.0",
     "@vitejs/plugin-vue": "^4.0.0",
     "@vue/eslint-config-typescript": "^12.0.0",
-    "@vue/test-utils": "^2.3.0",
     "autoprefixer": "^10.4.20",
-    "cypress": "^13.5.0",
-    "eslint": "^8.35.0",
-    "eslint-plugin-vue": "^9.9.0",
+    "cypress": "^13.15.1",
+    "eslint": "^8.57.1",
+    "eslint-plugin-vue": "^9.30.0",
+    "globals": "^15.11.0",
     "jsdom": "^22.1.0",
     "postcss": "^8.4.47",
+    "prettier": "^3.3.3",
     "tailwindcss": "^3.4.13",
     "terser": "^5.4.0",
     "typescript": "^5.1.6",
-    "vite": "~5.2.0",
+    "vite": "^5.4.10",
     "vitest": "^0.34.6",
     "vue-tsc": "^2.0.22"
   },
