@@ -1,9 +1,9 @@
-import type { CapacitorConfig } from '@capacitor/cli';
+import type { CapacitorConfig } from "@capacitor/cli";
 
 const config: CapacitorConfig = {
-  appId: 'io.ionic.starter',
-  appName: 'movil_user_mygardenllc',
-  webDir: 'dist'
+  appId: "io.ionic.starter",
+  appName: "movil_user_mygardenllc",
+  webDir: "dist",
 };
 
 export default config;
